24/8

+++ src/routers/admin/catergory/service.ts
@@ -6,6 +6,7 @@ import { CreateCategoryDto, GetAll, UpdateCategoryDto } from "./dto/dto";
 import { ByID, CodeParam, PaginationRes } from "src/interface/dto";
 import { FindAll, FindAllPagination, checkExit } from "src/utils";
 import { BadRequestException } from "@nestjs/common/exceptions";
+import { deleteFile } from "src/firebase";
 
 @Injectable()
 export class CategoryService {
@@ -62,6 +63,11 @@ export class CategoryService {
   }
 
   async remove({ id }: ByID): Promise<Category> {
+    const isExit = await checkExit(this.categoriesModel, {
+      _id: id,
+    });
+    if (!isExit) throw new BadRequestException("data wrong");
+    if (isExit.image && isExit.image != "") deleteFile(isExit.image);
     return this.categoriesModel.findByIdAndDelete(id);
   }
   async checkCode({ code }: CodeParam): Promise<boolean> {
